Changing so memory matches real life

+++ src/compiler/crystal/repl/instruction.cr
@@ -1,1 +1,1 @@
-alias Crystal::Repl::Instruction = Int64
+alias Crystal::Repl::Instruction = UInt8
