+++ src/compiler/crystal/repl/op_code.cr
@@ -2,7 +2,7 @@ require "./repl"
 require "./instructions"
 
 {% begin %}
-  enum Crystal::Repl::OpCode : Int64
+  enum Crystal::Repl::OpCode : UInt8
     {% for name, instruction in Crystal::Repl::Instructions %}
       {{ name.id.upcase }}
     {% end %}
